修改聊天界面消息都显示到一侧bug

+++ IMiPhone/Persistence/Database/DBUser.h
@@ -12,7 +12,7 @@
 
 @interface DBUser : NSObject
 
-@property (nonatomic) NSInteger uid;
+@property (nonatomic) long long uid;
 @property (nonatomic, retain) NSString *oid;
 @property (nonatomic, retain) NSString *nick;
 @property (nonatomic, retain) NSString *mobile;
